fix: unify casing to categoryChips
