+++ components/category/categoryChips.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import clsx from 'clsx';
-import React, { Dispatch, SetStateAction } from 'react';
+import React from 'react';
 
 import Art from '@/public/icons/category/art.svg';
 import IT from '@/public/icons/category/it.svg';
@@ -14,35 +14,44 @@ type Props = {
   categories?: string[];
   activeIndex?: number;
   className?: string;
-  onSelect: Dispatch<SetStateAction<number>>;
+  onSelect?: (idx: number) => void;
 
   boxSize?: number;
   iconPaddingRatio?: number;
 };
 
 type Cat = {
-  id: number;
   label: string;
   Icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
 };
 
 const FIXED_CATEGORIES: Cat[] = [
-  { id: 1, label: '창의 · 예술', Icon: Art },
-  { id: 2, label: '라이프스타일', Icon: Lifestyle },
-  { id: 3, label: '스포츠 · 웰빙', Icon: Sports },
-  { id: 4, label: '언어 · 학습', Icon: Language },
-  { id: 5, label: 'IT · 디지털', Icon: IT },
-  { id: 6, label: '자기계발', Icon: SelfDev },
+  { label: '창의 · 예술', Icon: Art },
+  { label: '라이프스타일', Icon: Lifestyle },
+  { label: '스포츠 · 웰빙', Icon: Sports },
+  { label: '언어 · 학습', Icon: Language },
+  { label: 'IT · 디지털', Icon: IT },
+  { label: '자기계발', Icon: SelfDev },
 ];
 
 const CATEGORY_VIEWBOX = '0 0 60 60';
 
-export default function CategoryChips({ activeIndex = 0, className = '', onSelect }: Props) {
+export default function CategoryChips({
+  categories,
+  activeIndex = 0,
+  className = '',
+  onSelect,
+  boxSize = 64,
+  iconPaddingRatio = 0.12,
+}: Props) {
+  const pad = Math.max(4, Math.round(boxSize * iconPaddingRatio));
+  const iconPx = Math.max(20, boxSize - pad * 2);
+
   return (
     <div className={clsx('overflow-x-auto scrollbar-hidden', className)}>
-      <ul className="flex gap-[2rem] pr-0">
-        {FIXED_CATEGORIES.map(({ id, label, Icon }) => {
-          const active = id === activeIndex;
+      <ul className="flex gap-[2.0rem] pl-[3.2rem] pr-0">
+        {FIXED_CATEGORIES.map(({ label, Icon }, idx) => {
+          const active = idx === activeIndex;
 
           return (
             <li
@@ -51,9 +60,9 @@ export default function CategoryChips({ activeIndex = 0, className = '', onSelec
             >
               <button
                 type="button"
-                onClick={() => onSelect(id)}
+                onClick={() => onSelect?.(idx)}
                 aria-pressed={active}
-                className="flex flex-col w-[8.2rem] items-center gap-[.4rem] focus:outline-none"
+                className="group flex w-[8.0rem] flex-col items-center gap-2 focus:outline-none"
               >
                 <span
                   className={clsx(
@@ -62,9 +71,14 @@ export default function CategoryChips({ activeIndex = 0, className = '', onSelec
                       ? 'bg-white border-[var(--black)]'
                       : 'bg-white border-[var(--gray1)] hover:border-[var(--black)]'
                   )}
+                  style={{ width: boxSize, height: boxSize }}
                 >
                   <Icon
+                    width={iconPx}
+                    height={iconPx}
                     viewBox={CATEGORY_VIEWBOX}
+                    preserveAspectRatio="xMidYMid meet"
+                    style={{ display: 'block' }}
                     aria-hidden
                     focusable="false"
                   />
@@ -82,6 +96,11 @@ export default function CategoryChips({ activeIndex = 0, className = '', onSelec
             </li>
           );
         })}
+
+        <li
+          aria-hidden
+          className="shrink-0 w-[3.2rem]"
+        />
       </ul>
     </div>
   );
